+++ src/main/kotlin/blackjack/game/GameEvaluator.kt
@@ -1,6 +1,5 @@
 package blackjack.game
 
-import blackjack.BlackjackGame
 import blackjack.dealer.Dealer
 import blackjack.player.Player
